feat: force video capture through command

+++ pinthesky/combiner.py
@@ -29,11 +29,14 @@ class VideoCombiner(Handler):
         with open(file_name, 'wb') as o:
             for n in ['before', 'after']:
                 part_name = f'{event["start_time"]}.{n}.h264'
+                if not os.path.exists(part_name):
+                    continue
                 with open(part_name, 'rb') as i:
                     o.write(i.read())
                 os.remove(part_name)
         self.events.fire_event('combine_end', {
             'start_time': event['start_time'],
-            'combine_video': file_name
+            'combine_video': file_name,
+            'trigger': event['trigger'],
         })
         logger.info(f'Finish concatinating to {file_name}')
